[WIP & Wallet] Add initial blockchain-backup base code
This is the base code for ZENZO's automatic blockchain backup system, this system intends to secure a wallet's local blockchain by backing it up to a seperate directory, if the main blockchain is damaged, the wallet will fallback to the backup directory, recovering a state that, at max, is only 15 to 20 blocks old. This feature is currently unfinished and only semi-functional.