[Qt] Add GUI-based debug log viewer
This is a GUI viewer of the wallet's debug.log file, allowing users to observe their logs with ease and without needing knowledge of their data directory or debug files. This will make it easier for users to report problems or bugs that can be observed via the debug.log output.